feat(agent): inline source dropdown in deposit modal, drop options screen
Mirror the card deposit modal's source selector. The agent deposit
modal is now a single screen with a 'From' dropdown at the top
(Borrow against Savings / External Wallet) and the matching form
rendered inline below. Removes the prior options-step indirection,
the back-button bookkeeping, and the per-option descriptions.

- Web: real DropdownMenu (radix) — same -mt-4 rounded-b-2xl pattern.
- Native: inline expanding section (Pressable + isOpen) — same as the
  card flow's SourceSelectorNative.
- AgentDepositBorrowForm extracted into its own file for symmetry with
  AgentDepositExternalForm.

(cherry picked from commit 106e1caaeeda7a92a08b228ad3ac2fbfc8e61dc9)
